+++ packages/plugin-packages/mashroom-security-provider-basic-wrapper/type-definitions/index.ts
@@ -1,0 +1,6 @@
+
+import {ExpressRequest} from '@mashroom/mashroom/type-definitions';
+
+export type ExpressRequestWithSession = ExpressRequest & {
+    session: any;
+}